refactor: rename package name, var name of encoder
