+++ plugins/extension/default_encoder/default_encoder.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package encoder
+package defaultencoder
 
 import (
 	"encoding/json"
